Create MountEXT4.ps1
This is a simple script to mount EXT4 drives via Windows Subsystem for Linux. It will not work unless you have WSL installed (https://learn.microsoft.com/en-us/windows/wsl/install)

This was originally created to mount my External EXT4 drive from my Steam Deck to Windows to access the games on it.

You'll access the drive similar to a network share. The path for the drive once mounted will be the following unless modified:

\\wsl.localhost\LINUXDISTROHERE\mnt\wsl\YOURDRIVEHERE